Use a countdown timer to schedule video frame captures.
This method avoids skew due to delay capturing the frame.
For example, without this change, if a frame takes 3ms to capture, and we delay
100ms between frames, then the existing code will collect
a frame at 103ms, 206ms, 309ms, and so on.  However, the
frames will be labled 100ms, 200ms, 300ms, giving the
user the impression that the page loads faster than it does.

BUG=
TEST=

Review URL: https://codereview.appspot.com/6118063